Enhancement: Allow passing a file-like object alongside with path and url in the factory method parse().
This pattern provides more flexibility to construct a variety of
file-like object such as BytesIO outside of the helper instead of
a regular file.

This commit also fixes a bug that a relative filesystem path starting
with "http", e.g., "http/boot.iso", will be opened as a http URL
incorrectly.